refactor: 코드리뷰 반영
1. NumberGenerator에 구현체가 의존하도록 타입 변경
2. 컨트롤러 패키지로 분리
