+++ src/main/java/racinggamewinner/random/NumberGenerateStrategy.java
@@ -1,15 +1,15 @@
 package racinggamewinner.random;
 
-public class NumberGenerateStrategy {
+public class NumberGenerateStrategy implements NumberGenerator {
     private NumberGenerator numberGeneratorStrategy;
 
     public NumberGenerateStrategy(NumberGenerator numberGeneratorStrategy) {
         this.numberGeneratorStrategy = numberGeneratorStrategy;
     }
 
+    @Override
     public int executeStrategy() {
-        return numberGeneratorStrategy.generateNumber();
+        return numberGeneratorStrategy.executeStrategy();
     }
-
 
 }
