+++ src/main/java/racinggamewinner/random/NumberGenerator.java
@@ -2,5 +2,5 @@ package racinggamewinner.random;
 
 @FunctionalInterface
 public interface NumberGenerator {
-    int generateNumber();
+    int executeStrategy();
 }
